add icon to footer, organize icon folders
