+++ src/components/Footer/Footer.module.scss
@@ -6,7 +6,7 @@
 	padding: 10rem 0 3.5rem;
 	font-size: 1.3rem;
 	margin-top: auto;
-	* {
+	&_portalInfo {
 		line-height: 1.8rem;
 	}
 	&_Logotype {
