+++ src/components/Icons/CloseIcon/CloseIcon.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/vars";
+@import "../../../styles/vars";
 
 .CloseIcon {
 	height: 2rem;
