+++ src/components/Icons/CloseIcon/CloseIcon.stories.tsx
