+++ src/components/Icons/CloseIcon/CloseIcon.tsx
