+++ src/components/Icons/MouseIcon.tsx
