+++ src/components/Icons/OrganoidIcon.tsx
