+++ src/components/Layout/Layout.tsx
@@ -6,9 +6,9 @@ import { QueryClient, QueryClientProvider } from "react-query";
 import { ReactQueryDevtools } from "react-query/devtools";
 import Button from "../Button/Button";
 import Card from "../Card/Card";
-import CloseIcon from "../CloseIcon/CloseIcon";
 import FeedbackIcon from "../FeedbackIcon/FeedbackIcon";
 import FloatingButton from "../FloatingWidget/FloatingButton";
+import CloseIcon from "../Icons/CloseIcon/CloseIcon";
 import Loader from "../Loader/Loader";
 import Modal from "../Modal/Modal";
 import Navbar from "../Navbar/Navbar";
